requirements.txt
Flask==3.0.0
requests==2.31.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
Flask-Scrypt==0.1.3.6
PyJWT==2.8.0
